split field definition into example/register_model.js and custom handler into example/controller.js
