+++ package.json
@@ -8,7 +8,8 @@
     "test": "test"
   },
   "scripts": {
-    "test": "tape"
+    "test": "tape",
+    "start":"node example/server.js"
   },
   "repository": {
     "type": "git",
